+++ lib/src/models/config/ck_configuration.dart
@@ -11,7 +11,9 @@ class CKConfiguration {
   final Set<CKCallAttributes> attributes;
   final CKTimer timer;
   final bool notifyOnCallEnd;
+  final bool notifyOnMissedCall;
   final Map<CKCallAction, String> icons;
+  final bool strictMode;
   // final bool repostOnClick;
 
   const CKConfiguration({
@@ -20,13 +22,15 @@ class CKConfiguration {
     this.attributes = const {},
     this.timer = const CKTimer(),
     this.notifyOnCallEnd = true,
+    this.notifyOnMissedCall = true,
     this.icons = const {},
+    this.strictMode = true,
     // this.repostOnClick = true,
   });
 
   @override
   String toString() {
     // return 'CKConfiguration{sounds: $sounds, capabilities: $capabilities, attributes: $attributes, timer: $timer, notifyOnCallEnd: $notifyOnCallEnd, repostOnClick: $repostOnClick}';
-    return 'CKConfiguration{sounds: $sounds, capabilities: $capabilities, attributes: $attributes, timer: $timer, notifyOnCallEnd: $notifyOnCallEnd, icons: $icons}';
+    return 'CKConfiguration{sounds: $sounds, capabilities: $capabilities, attributes: $attributes, timer: $timer, notifyOnCallEnd: $notifyOnCallEnd, icons: $icons, strictMode: $strictMode, notifyOnMissedCall: $notifyOnMissedCall}';
   }
 }