+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+* Add `strictMode` to `CKConfiguration`
+* Fix(web): Instance duplication on getting singleton in web implementation. 
+* update docs
+
 ## 0.0.3+1
 
 * BREAKING CHANGE: Use `WebCallKitWeb` instead of `WebCallKit`.